+++ CHANGELOG.md
@@ -2,6 +2,23 @@
 
 ## Next Release
 
+## 0.1.25
+
+- [sdk] sdk subgraph active polling
+
+## 0.1.24
+
+- [sdk] another hotfix for sdk subgraph polling
+
+## 0.1.23
+
+- [sdk] hotfix for sdk subgraph polling
+
+## 0.1.22
+
+- [sdk] type update for userAddress for sdkBase
+- [sdk] sdk polling only active txs
+- [utils] fix hardcoded gaslimit
 - [router] Remove relayerFeesPaid stuff when fulfilling on receiving side
 
 ## 0.1.21
@@ -35,6 +52,7 @@
 
 - [router] fix params for gas estimation before relayer use
 - [utils] Update hardcoded gasLimits for arbitrum network
+- [contracts] Add v1PriceOracle
 
 ## 0.1.14